chore: move SvgIcon out, testing stories
