+++ package.json
@@ -89,6 +89,7 @@
     "@types/react-dom": "^16.9.5",
     "@types/react-helmet": "5.0.13",
     "@types/react-redux": "7.1.5",
+    "@types/react-tooltip": "^3.11.0",
     "@typescript-eslint/eslint-plugin": "2.3.2",
     "@typescript-eslint/parser": "2.3.2",
     "babel-jest": "24.9.0",
