+++ src/components/AssetInfo/TokenSideBar.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import css from "./TokenSideBar.scss";
 import TokenSideBarContent from "./TokenSideBarContent";
 import TokenSideBarRole from "./TokenSideBarRole";
+import SvgIcon from "./../Helpers/SvgIcon";
 
 interface TokenSideBarContentProps {
   adminAddress: string;
@@ -46,20 +47,9 @@ const TokenSideBar = ({
         />
       </div>
       <div className={`${css.hamburger}`} onClick={handler}>
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          width="24"
-          height="24"
-          viewBox="0 0 24 24"
-          fill="none"
-          stroke="currentColor"
-          strokeWidth="2"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          className="feather feather-chevron-left"
-        >
+        <SvgIcon cssClass="feather-info">
           <polyline points="15 18 9 12 15 6" />
-        </svg>
+        </SvgIcon>
       </div>
     </aside>
   );
