+++ src/components/Helpers/SvgIcon.stories.mdx
@@ -1,0 +1,28 @@
+import { Meta, Story, Preview } from "@storybook/addon-docs/blocks";
+import SvgIcon from "./SvgIcon.tsx";
+
+<Meta title="Icons" component={SvgIcon} />
+
+# SvgIcon
+
+Icons used in application.
+
+<Preview>
+  <Story name="Icon Info">
+    <SvgIcon cssClass="feather-info">
+      <circle cx="12" cy="12" r="10" />
+      <line x1="12" y1="16" x2="12" y2="12" />
+      <line x1="12" y1="8" x2="12.01" y2="8" />
+    </SvgIcon>
+  </Story>
+  <Story name="Icon Check">
+    <SvgIcon cssClass="feather-check">
+      <polyline points="20 6 9 17 4 12" />
+    </SvgIcon>
+  </Story>
+  <Story name="Icon Chevron Left">
+    <SvgIcon cssClass="feather-chevron-left">
+      <polyline points="15 18 9 12 15 6" />
+    </SvgIcon>
+  </Story>
+</Preview>
